Move DirectClientManager and direct-connection factory to their own file
